// Presentation/cosmeticshop.client/src/apiClient/models/user-add-request-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CosmeticShop.WebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { RoleType } from './role-type';
 /**
 * 
 *
 * @export
 * @interface UserAddRequestDto
 */
export interface UserAddRequestDto {

    /**
     * @type {string}
     * @memberof UserAddRequestDto
     */
    firstName: string;

    /**
     * @type {string}
     * @memberof UserAddRequestDto
     */
    lastName: string;

    /**
     * @type {string}
     * @memberof UserAddRequestDto
     */
    email: string;

    /**
     * @type {string}
     * @memberof UserAddRequestDto
     */
    password: string;

    /**
     * @type {RoleType}
     * @memberof UserAddRequestDto
     */
    role: RoleType;
}