// Presentation/cosmeticshop.client/src/apiClient/models/user-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * CosmeticShop.WebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { RoleType } from './role-type';
 /**
 * 
 *
 * @export
 * @interface UserResponseDto
 */
export interface UserResponseDto {

    /**
     * @type {string}
     * @memberof UserResponseDto
     */
    token?: string | null;

    /**
     * @type {string}
     * @memberof UserResponseDto
     */
    id?: string;

    /**
     * @type {string}
     * @memberof UserResponseDto
     */
    firstName: string;

    /**
     * @type {string}
     * @memberof UserResponseDto
     */
    lastName: string;

    /**
     * @type {string}
     * @memberof UserResponseDto
     */
    email: string;

    /**
     * @type {RoleType}
     * @memberof UserResponseDto
     */
    role: RoleType;
}